% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client-parameters}
\alias{client-parameters}
\title{Parameters for client functions}
\arguments{
\item{client_id}{Client ID}

\item{concise}{Generate concise output}
}
\description{
Parameters for client functions
}
