% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client}
\alias{client}
\title{Get client}
\usage{
client(client_id, concise = TRUE)
}
\arguments{
\item{client_id}{Client ID}

\item{concise}{Generate concise output}
}
\value{
A data frame with one record per client
}
\description{
Get client
}
\examples{
\dontrun{
client("63a5493591ed63165538976d")
}
}
