% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_create}
\alias{client_create}
\title{Add a new client to workspace}
\usage{
client_create(name, concise = TRUE)
}
\arguments{
\item{name}{Client name}

\item{concise}{Generate concise output}
}
\value{
A data frame with one row per record.
}
\description{
Add a new client to workspace
}
\examples{
\dontrun{
client_create("RStudio")
}
}
