% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_delete}
\alias{client_delete}
\title{Delete a client from workspace}
\usage{
client_delete(client_id)
}
\arguments{
\item{client_id}{Client ID}
}
\value{
A Boolean: \code{TRUE} on success or \code{FALSE} on failure.
}
\description{
Delete a client from workspace
}
