% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_insert}
\alias{client_insert}
\title{Add a new client to workspace}
\usage{
client_insert(name, concise = TRUE)
}
\arguments{
\item{name}{Client name}

\item{concise}{Generate concise output}
}
\value{
A data frame with one row per record.
}
\description{
Wraps \code{POST /workspaces/{workspaceId}/clients}.
}
\examples{
\dontrun{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

client_insert("RStudio")
}
}
