% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_update}
\alias{client_update}
\title{Update a client}
\usage{
client_update(client_id, name = NULL, note = NULL, archived = NULL)
}
\arguments{
\item{client_id}{Client ID}

\item{name}{Client name}

\item{note}{Note about client}

\item{archived}{Whether or not client is archived}
}
\value{
A data frame with one record for the updated client.
}
\description{
Update a client
}
