% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{clients}
\alias{clients}
\title{Get clients}
\usage{
clients(concise = TRUE)
}
\arguments{
\item{concise}{Generate concise output}
}
\value{
A data frame with one record per client.
}
\description{
Get clients
}
\examples{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

clients()
}
