% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-fields.R
\name{custom_field_delete}
\alias{custom_field_delete}
\title{Remove a custom field from a project}
\usage{
custom_field_delete(project_id, custom_field_id)
}
\description{
Remove a custom field from a project
}
