% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{get_api_key}
\alias{get_api_key}
\title{Get API key}
\usage{
get_api_key()
}
\value{
The API key.
}
\description{
Get API key
}
\examples{
\dontrun{
CLOCKIFY_API_KEY <- Sys.getenv("CLOCKIFY_API_KEY")
set_api_key(CLOCKIFY_API_KEY)
get_api_key()
}
}
