% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{paginate}
\alias{paginate}
\title{Title}
\usage{
paginate(path, query = NULL, pages = NULL, page_size = 50)
}
\arguments{
\item{pages}{Maximum number of pages to retrieve.}
}
\value{

}
\description{
Title
}
