% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_burndown}
\alias{plot_burndown}
\title{Plot burndown}
\usage{
plot_burndown(d_times, beg, end, hrs)
}
\arguments{
\item{d_times}{data frame of times from \code{\link{time_entries_all}} with \code{duration_hrs} field}

\item{beg}{beginning date}

\item{end}{end date to draw red dashed line from beginning \code{beg} to end \code{end} starting with hours \code{hrs} to zero}

\item{hrs}{hours estimated for burn down from begin}
}
\value{
ggplot2 plot
}
\description{
Plot burndown
}
