% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{POST}
\alias{POST}
\title{POST file to an URL}
\usage{
POST(path, body = NULL)
}
\arguments{
\item{path}{The path of the endpoint.}

\item{body}{The body of the query.}
}
\value{
A \code{\link[httr:response]{response()}} object.
}
\description{
POST file to an URL
}
