% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project-update}
\alias{project-update}
\alias{project_update}
\alias{project_update_template}
\title{Update project}
\usage{
project_update(project_id, name = NULL, client_id = NULL, archived = NULL)

project_update_template(project_id, is_template = TRUE)
}
\arguments{
\item{project_id}{Project ID}

\item{name}{Project name}

\item{client_id}{Client ID}

\item{archived}{Whether or not item is archived}

\item{is_template}{Is project a template?}
}
\description{
Update project
}
