% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Get project}
\usage{
project(project_id, concise = TRUE)
}
\arguments{
\item{project_id}{Project ID}

\item{concise}{Generate concise output}
}
\value{
A data frame with one record per project
}
\description{
Get project
}
\examples{
\dontrun{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

project("612b16c0bc325f120a1e5099")
}
}
