% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_delete}
\alias{project_delete}
\title{Delete project}
\usage{
project_delete(project_id)
}
\arguments{
\item{project_id}{Project ID}
}
\description{
An active project cannot be deleted. Archive the project first.
}
