% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_update}
\alias{project_update}
\title{Update project}
\usage{
project_update(project_id, name = NULL, client_id = NULL, archived = NULL)
}
\arguments{
\item{project_id}{Project ID}

\item{name}{Project name}

\item{client_id}{Client ID}

\item{archived}{Whether or not item is archived}
}
\description{
Wraps \code{PUT /workspaces/{workspaceId}/projects/{projectId}}.
}
\examples{
\dontrun{
}
}
