% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_update_billable_rate}
\alias{project_update_billable_rate}
\title{Update user billable rate on project}
\usage{
project_update_billable_rate(project_id, user_id, rate, since = NULL)
}
\arguments{
\item{project_id}{Project ID}

\item{user_id}{User ID}

\item{rate}{Rate}

\item{since}{New rate will be applied to all time entries after this time}
}
\description{
Update user billable rate on project
}
