% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_update_memberships}
\alias{project_update_memberships}
\title{Update project memberships}
\usage{
project_update_memberships(project_id, user_id)
}
\arguments{
\item{project_id}{Project ID}

\item{user_id}{One or more user IDs}
}
\description{
Update project memberships
}
