% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{projects}
\alias{projects}
\title{Get projects}
\usage{
projects(concise = TRUE)
}
\arguments{
\item{concise}{Generate concise output}
}
\value{
A data frame with one record per project
}
\description{
Get projects
}
\examples{
\dontrun{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

projects()
}
}
