% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports_weekly}
\alias{reports_weekly}
\title{Weekly report}
\usage{
reports_weekly(start, end, extra_args = list(), tz = Sys.timezone())
}
\arguments{
\item{start}{Start time}

\item{end}{End time}

\item{extra_args}{Extra arguments to be passed to the
\href{https://docs.clockify.me/#tag/Time-Entry-Report/operation/generateSummaryReport}{API}.
Example: \code{extra_args = list(rounding = TRUE)}.}

\item{tz}{Time zone for output timestamps, default: Sys.timezone().}
}
\value{
A data frame with a weekly summary of time entries for the specified time period.
}
\description{
Weekly report
}
\examples{
\dontrun{
report <- reports_weekly("2022-08-01", "2022-08-08")

report \%>\%
  select(-duration, -amount) \%>\%
  unnest(projects)
}
}
