% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{A Clockify API key}
}
\value{
The API key.
}
\description{
Set API key
}
\examples{
CLOCKIFY_API_KEY <- Sys.getenv("CLOCKIFY_API_KEY")
set_api_key(CLOCKIFY_API_KEY)
}
