% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-shared.R
\name{shared_report}
\alias{shared_report}
\title{Get a specific shared report}
\usage{
shared_report(shared_report_id)
}
\arguments{
\item{shared_report_id}{Identifier for a specific shared report}
}
\description{
Get a specific shared report
}
\examples{
\dontrun{
# Get all shared reports.
shared_reports()
# Get specific shared report by shared report ID.
shared_report("6307f29f1bbd1d34e56b9eb7")
}
}
