% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-shared.R
\name{shared_report_delete}
\alias{shared_report_delete}
\title{Delete a shared report}
\usage{
shared_report_delete(shared_report_id)
}
\arguments{
\item{shared_report_id}{Identifier for a specific shared report}
}
\description{
Delete a shared report
}
\examples{
\dontrun{
shared_report_delete("6307f29f1bbd1d34e56b9eb7", name = "Test Report")
}
}
