% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag}
\alias{tag}
\title{Get tag}
\usage{
tag(tag_id)
}
\arguments{
\item{tag_id}{Tag ID}
}
\value{
A data frame with one record per tag
}
\description{
Get tag
}
\examples{
\dontrun{
tag("5f2d9bc659badb2a849c027e")
}
}
