% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_create}
\alias{tag_create}
\title{Create tag}
\usage{
tag_create(name)
}
\arguments{
\item{name}{Tag name}
}
\description{
Create tag
}
\examples{
\dontrun{
tag_create("Size: S")
tag_create("Size: M")
tag_create("Size: L")
tag_create("Size: XL")
}
}
