% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_update}
\alias{tag_update}
\title{Update tag}
\usage{
tag_update(tag_id, name = NULL, archived = NULL)
}
\arguments{
\item{tag_id}{Tag ID}

\item{name}{Tag name}

\item{archived}{Whether or not item is archived}
}
\description{
Update tag
}
\examples{
\dontrun{
tag_update("5f2d9bc659badb2a849c027e", "Size: Large")
tag_update("5f2d9bc659badb2a849c027e", archived = TRUE)
tag_update("5f2d9bc659badb2a849c027e", "Size: L", FALSE)
}
}
