% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{task-create}
\alias{task-create}
\alias{task_create}
\title{Create a task}
\usage{
task_create(project_id, name)
}
\arguments{
\item{project_id}{Project ID}

\item{name}{Task name}
}
\description{
Create a task
}
\examples{
\dontrun{
task_create("630ce53290cfd8789366fd49", "tests")
task_create("630ce53290cfd8789366fd49", "docs")
}
}
