% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{task_delete}
\alias{task_delete}
\title{Delete task}
\usage{
task_delete(project_id, task_id)
}
\arguments{
\item{project_id}{Project ID}

\item{task_id}{Task ID}
}
\description{
Delete task
}
\examples{
\dontrun{
task_delete("630ce53290cfd8789366fd49", "630ce57e25e863294e5c6cf2")
}
}
