% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{task_update}
\alias{task_update}
\title{Update a task}
\usage{
task_update(
  project_id,
  task_id,
  name,
  billable = NULL,
  status = NULL,
  assignee_id = NULL
)
}
\arguments{
\item{project_id}{Project ID}

\item{task_id}{Task ID}

\item{name}{Task name}

\item{billable}{Is the task billable?}

\item{status}{Is the task ACTIVE or DONE?}

\item{assignee_id}{Assignee ID}
}
\description{
Update a task
}
\examples{
\dontrun{
task_update("630ce53290cfd8789366fd49", "630ce57e25e863294e5c6cf2", "Tests")
task_create("630ce53290cfd8789366fd49", "630ce80a7f07da44c14ca9a2", "Docs", FALSE)
}
}
