% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{task_update_cost_rate}
\alias{task_update_cost_rate}
\title{Update task cost rate}
\usage{
task_update_cost_rate(project_id, task_id, rate, since = NULL)
}
\arguments{
\item{project_id}{Project ID}

\item{task_id}{Task ID}

\item{rate}{Rate}

\item{since}{New rate will be applied to all time entries after this time}
}
\description{
This feature is only available on the Pro and Enterprise plans.
}
