% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{tasks_all}
\alias{tasks_all}
\title{Get all tasks}
\usage{
tasks_all()
}
\value{
A data frame.
}
\description{
For all projects, wraps \code{GET /workspaces/{workspaceId}/projects/{projectId}/tasks}.
}
\examples{
\dontrun{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

tasks_all()
}
}
