% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entries}
\alias{time_entries}
\title{Title}
\usage{
time_entries(user_id, start = NULL, end = NULL, finished = TRUE, ...)
}
\arguments{
\item{finished}{Whether to include only finished time intervals (intervals with both start and end time).}

\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
# Specify number of results per page (default: 50).
\dontrun{
time_entries(user_id, page_size = 200)
}
# Specify number of pages.
\dontrun{
time_entries(user_id, pages = 3)
}
}
