% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entries_all}
\alias{time_entries_all}
\title{Get all time entries}
\usage{
time_entries_all()
}
\value{
A data frame with one record per time entry.
}
\description{
Get all time entries with expansion to all users and lookup names.
}
\examples{
\dontrun{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

USER_ID <- "612b15a4f4c3bf0462192676"

time_entries_all()
}
}
