% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entry}
\alias{time_entry}
\title{Insert a time entry}
\usage{
time_entry(project_id = NULL, start, end = NULL, description = NULL)
}
\arguments{
\item{project_id}{Project ID}

\item{start}{Start time}

\item{end}{End time}

\item{description}{Description}
}
\description{
You can only insert time entries for the authenticated used.
}
\examples{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

\dontrun{
time_entry(
  project_id = "600e73263e207962449a2c13",
  start = as.POSIXct("2021-01-02 08:00:00"),
  end   = as.POSIXct("2021-01-02 10:00:00"),
  description = "Doing stuff"
)
}
}
