% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entry_delete}
\alias{time_entry_delete}
\title{Delete a time entry}
\usage{
time_entry_delete(time_entry_id = NULL)
}
\arguments{
\item{time_entry_id}{Time entry ID}
}
\description{
Delete a time entry
}
