% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entry_invoiced}
\alias{time_entry_invoiced}
\title{Mark time entries as invoiced}
\usage{
time_entry_invoiced(time_entry_id, invoiced = TRUE)
}
\arguments{
\item{time_entry_id}{Time entry ID}

\item{invoiced}{Has this time entry been invoiced?}
}
\description{
The \code{time_entry_invoiced()} function will only work on a paid plan.
}
