% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entry_set}
\alias{time_entry_set}
\title{Replace a time entry}
\usage{
time_entry_set(
  time_entry_id,
  project_id = NULL,
  start,
  end = NULL,
  description = NULL
)
}
\arguments{
\item{time_entry_id}{Time entry ID}

\item{project_id}{Project ID}

\item{start}{Start time}

\item{end}{End time}

\item{description}{Description}
}
\description{
This does not update the time entry. It uses the same time entry ID but sets
all details from scratch.
}
