% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entry_stop}
\alias{time_entry_stop}
\title{Stop currently running timer}
\usage{
time_entry_stop(user_id = NULL, end = NULL)
}
\arguments{
\item{end}{End time}
}
\description{
Stop currently running timer
}
