% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time_format}
\alias{time_format}
\title{Convert time object into format required by API}
\usage{
time_format(time, to_utc = FALSE)
}
\arguments{
\item{time}{String representation of date and time}

\item{to_utc}{Whether to convert time to UTC}
}
\value{
A string representation of POSIXct.
}
\description{
Convert time object into format required by API
}
