% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time_parse}
\alias{time_parse}
\title{Title}
\usage{
time_parse(time, format = "\%Y-\%m-\%dT\%H:\%M:\%SZ")
}
\arguments{
\item{time}{}
}
\value{
A POSIXct object with UTC timezone.
}
\description{
Times returned by the API are all in UTC.
}
