% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user-update-role}
\alias{user-update-role}
\alias{user_update_role}
\title{Update user roles}
\usage{
user_update_role(user_id, role, entity_id)
}
\arguments{
\item{user_id}{User ID}

\item{role}{One of \code{"TEAM_MANAGER"}, \code{"PROJECT_MANAGER"} or
\code{"WORKSPACE_ADMIN"}.}

\item{entity_id}{Depending on \code{role}, this is a user ID (for
\code{"TEAM_MANAGER"}), a project ID (for \code{"PROJECT_MANAGER"}) or a workspace ID
(for \code{"WORKSPACE_ADMIN"}).}
}
\description{
Update user roles
}
