% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{Get information for logged in user}
\usage{
user(show_workspace = FALSE)
}
\arguments{
\item{show_workspace}{Show active and default workspace IDs}
}
\value{
A data frame with details of user profile.
}
\description{
Get information for logged in user
}
\examples{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

\dontrun{
user()
}
}
