% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_create}
\alias{user_create}
\title{Create a user}
\usage{
user_create(email, send_email = TRUE)
}
\arguments{
\item{email}{Email address for user}

\item{send_email}{Whether to send email to user}
}
\description{
Create a user
}
