% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-groups.R
\name{user_group_update}
\alias{user_group_update}
\title{Update a user group}
\usage{
user_group_update(group_id, name)
}
\arguments{
\item{group_id}{User group ID}

\item{name}{Name of user group}
}
\description{
Update a user group
}
