% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-groups.R
\name{user_group_user_add}
\alias{user_group_user_add}
\title{Add a user to a user group}
\usage{
user_group_user_add(group_id, user_id)
}
\arguments{
\item{group_id}{User group ID}

\item{user_id}{User ID}
}
\description{
Add a user to a user group
}
