% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-groups.R
\name{user_group_user_delete}
\alias{user_group_user_delete}
\title{Remove a user from a user group}
\usage{
user_group_user_delete(group_id, user_id)
}
\arguments{
\item{group_id}{User group ID}

\item{user_id}{User ID}
}
\description{
Remove a user from a user group
}
\examples{
\dontrun{
}
}
