% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-groups.R
\name{user_groups}
\alias{user_groups}
\title{Get user groups}
\usage{
user_groups()
}
\value{
A data frame with one record per user group.
}
\description{
Get user groups
}
\examples{
\dontrun{
user_groups()
}
}
