% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_update_hourly_rate}
\alias{user_update_hourly_rate}
\title{Update hourly rate}
\usage{
user_update_hourly_rate(user_id, rate, since = NULL)
}
\arguments{
\item{user_id}{User ID}

\item{rate}{Rate}

\item{since}{New rate will be applied to all time entries after this time}
}
\description{
Update hourly rate
}
