% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_update_status}
\alias{user_update_status}
\title{Update status}
\usage{
user_update_status(user_id, active)
}
\arguments{
\item{user_id}{User ID}

\item{active}{A Boolean indicating whether or not user is active. Can also specify either \code{"ACTIVE"} or \code{"INACTIVE"}.}
}
\description{
Update status
}
