% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspaces}
\alias{workspaces}
\title{Get a list of workspaces}
\usage{
workspaces()
}
\value{
A data frame with one record per workspace.
}
\description{
Get a list of workspaces
}
\examples{
\dontrun{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

workspaces()
}
}
