% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_delete}
\alias{client_delete}
\title{Delete a client from workspace}
\usage{
client_delete(client_id, archive = FALSE)
}
\arguments{
\item{client_id}{Client ID}

\item{archive}{Archive client before deleting.}
}
\value{
A Boolean: \code{TRUE} on success or \code{FALSE} on failure.
}
\description{
A client must first be archived before it can be deleted.
}
