% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-fields.R
\name{custom_field_update}
\alias{custom_field_update}
\title{Update a custom field on a project}
\usage{
custom_field_update(
  project_id,
  custom_field_id,
  default_value = NULL,
  status = NULL
)
}
\arguments{
\item{project_id}{Project ID}

\item{custom_field_id}{Custom field ID}

\item{default_value}{A default value for the field}

\item{status}{Status}
}
\description{
Update a custom field on a project
}
