% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-fields.R
\name{custom_fields}
\alias{custom_fields}
\title{Get custom fields}
\usage{
custom_fields(project_id = NULL)
}
\arguments{
\item{project_id}{Project ID}
}
\description{
Custom fields are only listed for specific projects if the default values for
those fields have been modified for those projects.
}
\details{
Custom fields are only available on the Pro and Enterprise plans.
}
