% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{GET}
\alias{GET}
\title{GET an URL}
\usage{
GET(path, query = NULL)
}
\arguments{
\item{path}{The path of the endpoint.}

\item{query}{The query parameters.}
}
\value{
A \code{\link[httr:response]{response()}} object.
}
\description{
GET an URL
}
