% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paginate.R
\name{paginate}
\alias{paginate}
\title{Title}
\usage{
paginate(path, query = NULL, pages = NULL, page_size = 50)
}
\arguments{
\item{path}{The path of the endpoint.}

\item{query}{The query parameters.}

\item{pages}{Maximum number of pages to retrieve.}

\item{page_size}{Number of results requested per page.}
}
\value{
Paginated response from API.
}
\description{
Title
}
