% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project-update}
\alias{project-update}
\alias{project_update}
\alias{project_update_template}
\title{Update project}
\usage{
project_update(project_id, name = NULL, client_id = NULL, archived = NULL)

project_update_template(project_id, is_template = TRUE)
}
\arguments{
\item{project_id}{Project ID}

\item{name}{Project name}

\item{client_id}{Client ID}

\item{archived}{Whether or not project is archived}

\item{is_template}{Whether or not project is a template}
}
\description{
Adjust the project characteristics.
}
\details{
These functions enable the following functionality:
\itemize{
\item change the project name
\item change the client ID associated with the project
\item toggle whether project is archived and
\item toggle whether project is a template (paid plan only).
}
}
