% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports-parameters}
\alias{reports-parameters}
\title{Reports Parameters}
\arguments{
\item{start}{Start time}

\item{end}{End time}

\item{extra_args}{Extra arguments to be passed to the \href{https://docs.clockify.me/#tag/Time-Entry-Report}{API}.}
}
\description{
These are parameters which occur commonly across functions for reports.
}
