% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports-parameters}
\alias{reports-parameters}
\title{Reports Parameters}
\arguments{
\item{start}{Start time}

\item{end}{End time}
}
\description{
These are parameters which occur commonly across functions for reports.
}
