% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports_detailed}
\alias{reports_detailed}
\title{Detailed report}
\usage{
reports_detailed(start, end)
}
\arguments{
\item{start}{Start time}

\item{end}{End time}
}
\value{
A data frame with detailed time entries for the specified time period.
}
\description{
Detailed report
}
\examples{
\dontrun{
report <- reports_detailed("2022-08-01", "2022-09-01")
}
}
