% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{reports_summary}
\alias{reports_summary}
\title{Summary report}
\usage{
reports_summary(start, end)
}
\arguments{
\item{start}{Start time}

\item{end}{End time}
}
\value{
A list with two elements, \code{totals} and \code{details}, each of which is a
data frame.
}
\description{
Summary report
}
\examples{
\dontrun{
report <- reports_summary("2022-08-01", "2022-09-01")

# Summary per user.
report$details
# Summary per client/project.
report$details \%>\%
  select(-duration, -amount, -amounts) \%>\%
  unnest(children)
# Summary per time entry.
report$details \%>\%
  select(-duration, -amount, -amounts) \%>\%
  unnest(children) \%>\%
  select(-duration, -amount) \%>\%
  unnest(children)
}
}
