% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-shared.R
\name{shared_report_update}
\alias{shared_report_update}
\title{Update a shared report}
\usage{
shared_report_update(
  shared_report_id,
  name = NULL,
  is_public = NULL,
  fixed_date = NULL
)
}
\arguments{
\item{shared_report_id}{Identifier for a specific shared report}

\item{name}{Report name}

\item{is_public}{Is this a public report?}

\item{fixed_date}{Are the dates fixed?}
}
\description{
Update a shared report
}
\examples{
\dontrun{
shared_report_update("6307f29f1bbd1d34e56b9eb7", name = "Test Report")
}
}
