% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entry}
\alias{time_entry}
\title{Get a specific time entry on workspace}
\usage{
time_entry(time_entry_id, concise = TRUE)
}
\arguments{
\item{time_entry_id}{Time entry ID}
}
\value{
A data frame with one record per time entry.
}
\description{
Wraps \code{GET /workspaces/{workspaceId}/time-entries/{id}}.
}
