% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entry}
\alias{time_entry}
\title{Insert a Time Entry}
\usage{
time_entry(
  workspace_id,
  project_id = NULL,
  user_id = NULL,
  start,
  end = NULL,
  description = NULL,
  billable = NULL
)
}
\arguments{
\item{billable}{}
}
\value{

}
\description{
Insert a Time Entry
}
\examples{
time_entry(
  workspace_id = "5ef23294df73064140f60bfc",
  project_id = "600e73263e207962449a2c13",
  start = as.POSIXct("2021-01-02 08:00:00"),
  end   = as.POSIXct("2021-01-02 10:00:00"),
  description = "Doing stuff"
)
}
