% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entry}
\alias{time_entry}
\title{Get a specific time entry on workspace}
\usage{
time_entry(time_entry_id, concise = TRUE)
}
\arguments{
\item{time_entry_id}{Time entry ID}

\item{concise}{Generate concise output}
}
\value{
A data frame with one record per time entry.
}
\description{
Get a specific time entry on workspace
}
\examples{
\dontrun{
time_entry("61343d27ab05e02be2c8c266")
}
}
