% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-entry.R
\name{time_entry_delete}
\alias{time_entry_delete}
\title{Delete a time entry}
\usage{
time_entry_delete(time_entry_id = NULL)
}
\arguments{
\item{time_entry_id}{Time entry ID}
}
\value{
A Boolean: \code{TRUE} on success or \code{FALSE} on failure.
}
\description{
Delete a time entry
}
\examples{
\dontrun{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

time_entry_delete("612c7bd2a34530476ab25c67")
}
}
