% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{users}
\alias{users}
\title{Get list of users in active workspace}
\usage{
users(active = TRUE, show_workspace = FALSE)
}
\arguments{
\item{active}{Only include active users}

\item{show_workspace}{Show active and default workspace IDs}
}
\value{
A data frame with one record per user.
}
\description{
Get list of users in active workspace
}
