% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{users}
\alias{users}
\title{Get list of users in active workspace}
\usage{
users(active = NULL, concise = TRUE)
}
\arguments{
\item{active}{Only include active users}

\item{concise}{Generate concise output}
}
\value{
A data frame with one record per user.
}
\description{
Get list of users in active workspace
}
\examples{
\dontrun{
set_api_key(Sys.getenv("CLOCKIFY_API_KEY"))

# Show only active users.
users()
# Show all users.
users(active = FALSE)
# Show active & default workspace for each user.
users(concise = FALSE)
}
}
